AI-Daily-Digest/requirements.txt
arxiv>=2.1.3,<3
gtrending>=0.3.6,<1
Jinja2>=3.1.4,<4
openai>=1.51.0,<2
PyGithub>=2.5.0,<3
PyYAML>=6.0.2,<7
requests>=2.32.3,<3
